Moved SnapshotMapLegend styled-components to SnapshotMapLegend/styles.js
