+++ src/components/SnapshotMapLegend/styles.js
